create fatty acid module
