+++ src/modules/index.ts
@@ -3,3 +3,4 @@ export * from "./food";
 export * from "./main";
 export * from "./unit";
 export * from "./aminoAcid";
+export * from "./fattyAcid";
